clean warning

+++ DmToolsApp/Features/Library/Spell/LibrarySpellViewModel.cs
@@ -67,6 +67,9 @@ namespace DmToolsApp.Features.Library
 
             var item = SelectedSpellItem;
 
+            if (!await ConfirmDeleteAsync(item.Title))
+                return;
+
             await _libraryDataService.DeleteLibraryItem(item);
 
             SpellItems.Remove(item);
